fix terser

+++ tools/build.sh
@@ -19,7 +19,10 @@ set_version() {
 
 uglify() {
     pushd $DEST/src/scripts
-    terser --ecma 8 -o seed_.js seed.js || echo "Build failed, terser is missing. install with \`npm install -g terser\`"; exit 1
+    if ! terser --ecma 8 -o seed_.js seed.js; then
+        echo "Build failed, terser is missing. install with \`npm install -g terser\`";
+        exit 1;
+    fi
     sed -i -e "s/{SEED_CONTENTS}/$(sed 's:[/\\&]:\\&:g' seed_.js)/" content_script.js
     rm -f seed*.js;
 }
